requirements.txt
tensorflow
opencv-python
numpy 
pillow 
keyboard 
pygetwindow
jupyter